\name{coda-discrete-utils-package}
\alias{coda-discrete-utils-package}
\alias{coda-discrete-utils}
\docType{package}
\title{
Coda-discrete-utils
}
\description{
Plots PMF and CDF from coda for discrete models
}
\details{
\tabular{ll}{
Package: \tab coda-discrete-utils\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-20\cr
License: \tab GPL v2.1\cr
}
Plots PMF and CDF from coda for discrete models.
}
\author{
Maintainer: Alexandre Bouchard <bouchard@stat.ubc.ca>
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{
}
