% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coda.cdf}
\alias{coda.cdf}
\title{Plot an approximate CDF (cumulative distribution function) from the output of rjags on
a discrete model.}
\usage{
coda.cdf(jags.output)
}
\arguments{
\item{jags.output}{The output of jags.sample.}
}
\description{
Plot an approximate CDF (cumulative distribution function) from the output of rjags on
a discrete model.
}

