% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coda.pmf}
\alias{coda.pmf}
\title{Plot an approximate PMF (probability mass function) from the output of rjags on
a discrete model.}
\usage{
coda.pmf(jags.output)
}
\arguments{
\item{jags.output}{The output of jags.sample.}
}
\description{
Plot an approximate PMF (probability mass function) from the output of rjags on
a discrete model.
}

